refactor(flags): one ctx key and one transport for outbound rule flags
The Type-2 (context-passed) rule flags each grew their own support layer
in internal/client: custom_user_agent had its own transport mounted by two
constructors, extra_headers had another transport smuggled in through
wrapWithLogging, context_1m was read in six SDK methods, claude_org_id in
a constructor — four flags, four mechanisms, each with its own context key
and applyXxx helper.

Rule flags are one design, so give them one data path and one transport:

- typ: a single RuleFlagsKey carries the resolved RuleFlags struct
  (WithRuleFlags/GetRuleFlags). The custom_user_agent / extra_headers /
  context_1m / claude_org_id keys and their With/Get pairs are gone.
  ClientUserAgentKey stays — the inbound client UA is a request fact, not
  a rule flag.
- protocolserver: ResolveRuleFlagsWithScenario attaches the whole resolved
  flag set once (applyRuleFlags), replacing the five per-flag apply
  helpers.
- client: ruleFlagTransport merges userAgentTransport +
  extraHeadersTransport. Per-chain applicability is a plain judgment:
  extra_headers apply on api_key providers only, UA resolution only on
  chains mounted with resolveUA (generic OpenAI / non-OAuth Anthropic;
  the generic Google chain mounts without it). Vendor chains (Claude
  OAuth, Codex, Kimi, Gemini, Antigravity) mount nothing, which is what
  keeps their pinned handshake headers decisive. wrapWithLogging is
  logging-only again.
- context_1m stays at the SDK layer (Betas field / per-call header
  option) on purpose — it is the one flag that must reach the Claude
  OAuth vendor chain — and claude_org_id stays at client construction;
  both now read the same GetRuleFlags.

No behavior change; the merged transport preserves the old two-transport
precedence (extra headers first, UA resolution wins on conflict, inner
vendor pins win over both).
